--- setup.py
@@ -1,12 +1,0 @@
-from distutils.core import setup
-
-setup(
-    name='capiq-python',
-    version='',
-    packages=['capiq_client', 'capiq_client.capiq_client.tests', 'capiq_client.capiq_client.tests.unit'],
-    url='',
-    license='',
-    author='gmann',
-    author_email='',
-    description=''
-)
