+++ tests/unit/test_capiq_client_gdsg.py
@@ -1,12 +1,16 @@
-from mock import mock
 import unittest
-from capiq_client.capiq_client import CapIQClient
 
-def mocked_gdst_data_requests_post(*args, **kwargs):
+from mock import mock
+
+from capiq_client import CapIQClient
+
+
+def mocked_gdsg_data_requests_post(*args, **kwargs):
     class MockResponse:
         def __init__(self, json_data, status_code):
             self.json_data = json_data
             self.status_code = status_code
+
         def json(self):
             return self.json_data
     """
@@ -21,7 +25,7 @@ def mocked_gdst_data_requests_post(*args, **kwargs):
             "NumCols": 1,
             "Seniority": "",
             "Mnemonic": "IQ_CLOSEPRICE",
-            "Function": "GDST",
+            "Function": "GDSG",
             "ErrMsg": None,
             "Properties": {},
             "StartDate": "",
@@ -33,7 +37,7 @@ def mocked_gdst_data_requests_post(*args, **kwargs):
             "Limit": ""
         }]}, 200)
 
-def mocked_gdst_no_data_requests_post(*args, **kwargs):
+def mocked_gdsg_no_data_requests_post(*args, **kwargs):
     class MockResponse:
         def __init__(self, json_data, status_code):
             self.json_data = json_data
@@ -41,11 +45,7 @@ def mocked_gdst_no_data_requests_post(*args, **kwargs):
 
         def json(self):
             return self.json_data
-    """
-    if args[0] == 'http://someurl.com/test.json':
-        return MockResponse({"key1": "value1"}, 200)
-    elif args[0] == 'http://someotherurl.com/anothertest.json':
-    """
+
     if args[0] is not None:
         return MockResponse(
             {
@@ -57,7 +57,7 @@ def mocked_gdst_no_data_requests_post(*args, **kwargs):
                             "NumCols": 1,
                             "Seniority": "",
                             "Mnemonic": "IQ_CLOSEPRICE",
-                            "Function": "GDST",
+                            "Function": "GDSG",
                             "ErrMsg": "SOME ERROR",
                             "Properties": {},
                             "StartDate": "",
@@ -71,29 +71,16 @@ def mocked_gdst_no_data_requests_post(*args, **kwargs):
                     ]
             }, 200)
 
-class TestCapiqClientGdst(unittest.TestCase):
+class TestCapiqClientGdsg(unittest.TestCase):
 
-    @mock.patch('capiq_client.capiq_client.requests.post', side_effect=mocked_gdst_data_requests_post)
-    def test_gdst_data(self, mocked_post):
+    @mock.patch('capiq_client.requests.post', side_effect=mocked_gdsg_data_requests_post)
+    def test_gdsg_data(self, mocked_post):
         ciq_client = CapIQClient()
-        return_value = ciq_client.gdst(["TRIP"], ["IQ_CLOSEPRICE"], ["close_price"], properties=[{}])
+        return_value = ciq_client.gdsg(["TRIP"], ["IQ_CLOSEPRICE"], ["close_price"], properties=[{}])
         self.assertEqual(return_value, {'TRIP:': {'close_price': '46.80'}})
 
-    @mock.patch('capiq_client.capiq_client.requests.post', side_effect=mocked_gdst_data_requests_post)
-    def test_gdst_data_no_properties(self, mocked_post):
+    @mock.patch('capiq_client.requests.post', side_effect=mocked_gdsg_no_data_requests_post)
+    def test_gdsg_no_data(self, mocked_post):
         ciq_client = CapIQClient()
-        return_value = ciq_client.gdst(
-            ["TRIP"],
-            ["IQ_CLOSEPRICE"],
-            ["close_price"],
-            start_date="12/19/1980",
-            end_date="12/19/2000",
-            frequency="M"
-        )
-        self.assertEqual(return_value, {'TRIP:': {'close_price': '46.80'}})
-
-    @mock.patch('capiq_client.capiq_client.requests.post', side_effect=mocked_gdst_no_data_requests_post)
-    def test_gdst_no_data(self, mocked_post):
-        ciq_client = CapIQClient()
-        return_value = ciq_client.gdst(["TRIP"], ["IQ_CLOSEPRICE"], ["close_price"], [{}])
+        return_value = ciq_client.gdsg(["TRIP"], ["IQ_CLOSEPRICE"], ["close_price"], [{}])
         self.assertEqual(return_value, {'TRIP:': {'close_price': None}})
