+++ tests/unit/test_capiq_client_gdsp.py
@@ -1,6 +1,9 @@
-from mock import mock
 import unittest
-from capiq_client.capiq_client import CapIQClient
+
+from mock import mock
+
+from capiq_client import CapIQClient
+
 
 def mocked_gdsp_data_requests_post(*args, **kwargs):
     class MockResponse:
@@ -70,13 +73,13 @@ def mocked_gdsp_no_data_requests_post(*args, **kwargs):
 
 class TestCapiqClientGdsp(unittest.TestCase):
 
-    @mock.patch('capiq_client.capiq_client.requests.post', side_effect=mocked_gdsp_data_requests_post)
+    @mock.patch('capiq_client.requests.post', side_effect=mocked_gdsp_data_requests_post)
     def test_gdsp_data(self, mocked_post):
         ciq_client = CapIQClient()
         return_value = ciq_client.gdsp(["TRIP"], ["IQ_CLOSEPRICE"], ["close_price"], [{}])
         self.assertEqual(return_value, {'TRIP:': {'close_price': '46.80'}})
 
-    @mock.patch('capiq_client.capiq_client.requests.post', side_effect=mocked_gdsp_no_data_requests_post)
+    @mock.patch('capiq_client.requests.post', side_effect=mocked_gdsp_no_data_requests_post)
     def test_gdsp_no_data(self, mocked_post):
         ciq_client = CapIQClient()
         return_value = ciq_client.gdsp(["TRIP"], ["IQ_CLOSEPRICE"], ["close_price"], [{}])
